Some more networking and devices work
The Kernel now has a list of Devices and also created a way to potentially support multiple networking devices.
Then also started work on the Networking stack, which for now is a simpe Buffer type and parts of the Ethernet Packet, which allows you to
decode a buffer and get its src+dest mac addresses as well as the payload. This allows you to discard mismatched packets early on.

There still needs some more thought into how the Networking should work in general, as I dont really have a good Vision for how it should
work and be structured yet