Start of major changes
Started work on a rather large rewrite to turn it more into a microkernel architecture to make development and configuration easier.

To start this, I removed the kernel from the current OS binary itself and moved it into its own crate to more clearly seperate it
and dont mix parts up and also allow for easier development on the kernel itself.
I then slowly started transitioning the OS binary over to using the new Kernel and having the old Kernel only serve as an interface
to the currently supported hardware and slowly remove more and more of its logic.