Decoupled more networking stuff
Did some more work on the networking stuff to decouple the different Parts, which actually allowed me to remove some statics and therefore locks
as the state was no longer shared. This obviously also helps with preventing deadlocks and in general makes it easier to think
about all things considered.