Worked on my custom WASM vm
Worked some more on the WASM parser stuff and also started work on a WASM vm.

Currently the vm will just interpret the WASM instruction by instruction and emulate the correct behaviour,
external functions can be provided by a simple Handler trait, which is designed to be async. This design allows
you to easily perform whatever the OS needs to do on a syscall and even swap it out and all that stuff.
We will see how well this sort of design translates