First basic WASM integration
Started with slowly integrating the WASM stuff into the kernel. this now allows you to run very basic
WASM programs, basically only i32 arithemtic.
Next up we need to improve the safety, as currently there are a couple of cases in which the WASM-vm will
panic, which is very undesired and should instead report the appropriate errors