A lot more improvements
Improved the WASM interpreter again in an effort to make it actually correct as the current behaviour is not
the correct one from what i am seeing.
Then also switched the memory implementation for the WASM extension to using a custom PageTable instead of a Vec,
which should improve performance and also not use up a large space on the Heap, which was rather wasteful before.

Then also started writing more documentation for the OS as this has been seriously lacking but especially now with
things getting more complicated it is definetly needed at least in the critical parts